Style embedding fixed
Starting from the bottom and working up, each level should start with a string consisting of its ANSI styling code preceded by the minimal reset code `\u{1b}[22;23;24m` to avoid any outer bold / dim / italic / underlined attributes bleeding through, and replace any internal reset `\u{1b}[0m` with the same string to reset attributes and re-establish its style after an embed. Finally it should append  a reset of its own (`\u{1b}[0m`), which will be left in as a reset if not embedded, or replaced in similar fashion if embedded in another string.
In this way, each string is only touched by the level immediately above it, and only to replace its final reset with the minimal reset and the enclosing level's ANSI code.
Note that `\u{1b}[`, `\x1b[`, `\033[` and `\e[` are alternative notations for the ANSI escape code.